Modified BioXASBeamStatus to take on the functionality of both BioXASSideBeamStatus and BioXASMainBeamStatus: reports the beam status as a function of shutters, valves, mono mask, mirror mask, and endstation shutter. To do this, modified BioXASShutters to take on the functionality of both BioXASSideShutters and BioXASMainShutters: reports the shutters status as a function of the front-end shutters and the endstation shutter. Created new class BioXASShuttersGroup that has the functionality previously implemented in BioXASShutters, a general interface for adding and removing shutters from a group. BioXASShutters and BioXASFrontEndShutters now both inherit from this new class. Generalized BioXASBeamStatusBar and BioXASBeamStatusView classes. Created instance of BioXASBeamStatusView in BioXASAppController, the view now appears in both Side and Main Acquaman applications.
