+++ BioXASSideAcquaman_internal.pro
@@ -6,9 +6,3 @@ TARGET = BioXASSideAcquaman
 
 SOURCES += \
 	source/application/BioXAS/BioXASSideMain.cpp
-
-HEADERS +=
-
-
-
-
