+++ bioxasMainCommon.pri
@@ -19,9 +19,7 @@ HEADERS += \
     source/beamline/BioXAS/BioXASMainDBHRMirrors.h \
 	source/beamline/BioXAS/BioXASMainM2MirrorBendControl.h \
 	source/beamline/BioXAS/BioXASMainStandardsWheel.h \
-	source/beamline/BioXAS/BioXASMainMonochromatorMask.h \
-    source/beamline/BioXAS/BioXASMainShutters.h \
-    source/beamline/BioXAS/BioXASMainBeamStatus.h
+	source/beamline/BioXAS/BioXASMainMonochromatorMask.h
 
 SOURCES += \
     source/application/BioXAS/BioXASMainAppController.cpp \
@@ -36,9 +34,7 @@ SOURCES += \
     source/beamline/BioXAS/BioXASMainDBHRMirrors.cpp \
 	source/beamline/BioXAS/BioXASMainM2MirrorBendControl.cpp \
 	source/beamline/BioXAS/BioXASMainStandardsWheel.cpp \
-	source/beamline/BioXAS/BioXASMainMonochromatorMask.cpp \
-    source/beamline/BioXAS/BioXASMainShutters.cpp \
-    source/beamline/BioXAS/BioXASMainBeamStatus.cpp
+	source/beamline/BioXAS/BioXASMainMonochromatorMask.cpp
