+++ source/application/BioXAS/BioXASSideAppController.cpp
@@ -23,7 +23,6 @@ along with Acquaman.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "beamline/BioXAS/BioXASSideBeamline.h"
 #include "ui/BioXAS/BioXASSidePersistentView.h"
-#include "ui/BioXAS/BioXASSideBeamStatusView.h"
 
 BioXASSideAppController::BioXASSideAppController(QObject *parent)
 	: BioXASAppController(parent)
@@ -75,31 +74,6 @@ void BioXASSideAppController::setupUserInterface()
 bool BioXASSideAppController::setupDataFolder()
 {
 	return AMChooseDataFolderDialog::getDataFolder("/AcquamanLocalData/bioxas-s/AcquamanSideData", "/home/bioxas-s/AcquamanSideData", "users", QStringList());
-}
-
-QWidget* BioXASSideAppController::createComponentView(QObject *component)
-{
-	QWidget *componentView = 0;
-
-	if (component) {
-		bool componentFound = false;
-
-		// Try to match up given component with known component types.
-		// If match found, create appropriate view.
-
-		BioXASSideBeamStatus *beamStatus = qobject_cast<BioXASSideBeamStatus*>(component);
-		if (!componentFound && beamStatus) {
-			componentView = new BioXASSideBeamStatusView(beamStatus);
-			componentFound = true;
-		}
-
-		// Finally, check to see if there is a general view for the given component.
-
-		if (!componentFound)
-			componentView = BioXASAppController::createComponentView(component);
-	}
-
-	return componentView;
 }
 
 void BioXASSideAppController::setupXASScanConfiguration(BioXASXASScanConfiguration *configuration)
