+++ source/icons/icons.qrc
@@ -218,9 +218,6 @@
         <file>shutterIcon2.png</file>
         <file>valveIcon2.png</file>
         <file>mono-icon.png</file>
-        <file>mono-icon2.png</file>
-        <file>mono-icon3.png</file>
-        <file>mono-icon4.png</file>
         <file>mono-icon5.png</file>
         <file>beam-icon.png</file>
         <file>mirror-icon1.png</file>
