+++ source/ui/BioXAS/BioXASBeamStatusView.cpp
@@ -1,6 +1,7 @@
 #include "BioXASBeamStatusView.h"
 #include "beamline/BioXAS/BioXASBeamStatus.h"
 #include "ui/BioXAS/BioXASControlEditor.h"
+#include "ui/BioXAS/BioXASBeamStatusBar.h"
 
 BioXASBeamStatusView::BioXASBeamStatusView(BioXASBeamStatus *beamStatus, QWidget *parent) :
     QWidget(parent)
@@ -12,18 +13,29 @@ BioXASBeamStatusView::BioXASBeamStatusView(BioXASBeamStatus *beamStatus, QWidget
 	// Create UI elements.
 
 	beamStatusEditor_ = new BioXASControlEditor(0);
+	beamStatusEditor_->setTitle("Beam status");
+
+	statusBar_ = new BioXASBeamStatusBar(0);
+
+	QVBoxLayout *statusBarLayout = new QVBoxLayout();
+	statusBarLayout->addWidget(statusBar_);
+
+	QGroupBox *statusBox = new QGroupBox();
+	statusBox->setTitle("Components");
+	statusBox->setLayout(statusBarLayout);
 
 	// Create and set layouts.
 
 	QVBoxLayout *layout = new QVBoxLayout();
+	layout->setMargin(0);
 	layout->addWidget(beamStatusEditor_);
+	layout->addWidget(statusBox);
 
 	setLayout(layout);
 
 	// Current settings.
 
 	setBeamStatus(beamStatus);
-
 	refresh();
 }
 
@@ -35,6 +47,7 @@ BioXASBeamStatusView::~BioXASBeamStatusView()
 void BioXASBeamStatusView::refresh()
 {
 	updateBeamStatusEditor();
+	updateBeamStatusBar();
 }
 
 void BioXASBeamStatusView::setBeamStatus(BioXASBeamStatus *newStatus)
@@ -51,4 +64,9 @@ void BioXASBeamStatusView::setBeamStatus(BioXASBeamStatus *newStatus)
 void BioXASBeamStatusView::updateBeamStatusEditor()
 {
 	beamStatusEditor_->setControl(beamStatus_);
+}
+
+void BioXASBeamStatusView::updateBeamStatusBar()
+{
+	statusBar_->setBeamStatus(beamStatus_);
 }
